fix: reduce global task search scope
