+++ apps/desktop/src/api/apiService.ts
@@ -137,7 +137,7 @@ export interface ApiService {
     direction?: TaskDependencyDirection,
   ): Promise<TaskDependencyListResponse>;
   listTasks(input: TaskListInput): Promise<TaskListPage>;
-  searchTasks(input: TaskSearchInput, signal?: AbortSignal): Promise<TaskSearchResponse>;
+  searchTasks(input: TaskSearchInput): Promise<TaskSearchResponse>;
   updateTask(input: TaskEditInput): Promise<string>;
   startTask(input: TaskStartInput): Promise<TaskStartResponse>;
   moveTask(input: TaskMoveInput): Promise<TaskMoveResponse>;
