+++ apps/desktop/src/api/client.ts
@@ -532,8 +532,8 @@ export class ApiClient implements ApiService {
     return workflowLabels.listTasks(this.#transport, input);
   }
 
-  async searchTasks(input: TaskSearchInput, signal?: AbortSignal): Promise<TaskSearchResponse> {
-    return taskSearch.searchTasks(this.#transport, input, signal);
+  async searchTasks(input: TaskSearchInput): Promise<TaskSearchResponse> {
+    return taskSearch.searchTasks(this.#transport, input);
   }
 
   async updateTask(input: TaskEditInput): Promise<string> {
