+++ apps/desktop/src/api/clientTaskSearch.test.ts
@@ -73,9 +73,7 @@ describe("ApiClient task search", () => {
   it("sends the exact literal search payload and maps the grouped response", async () => {
     const transport = new FakeRpcTransport([{ method: "workflow.task.search", result: literalResponse }]);
     const client = new ApiClient(transport);
-    const controller = new AbortController();
-
-    await expect(client.searchTasks(literalInput, controller.signal)).resolves.toEqual({
+    await expect(client.searchTasks(literalInput)).resolves.toEqual({
       mode: "literal",
       groups: [
         {
@@ -121,8 +119,7 @@ describe("ApiClient task search", () => {
       nextOffset: 25,
     });
 
-    expect(transport.calls).toEqual([]);
-    expect(transport.dedicatedCalls).toEqual([
+    expect(transport.calls).toEqual([
       {
         method: "workflow.task.search",
         params: {
@@ -135,7 +132,6 @@ describe("ApiClient task search", () => {
           page_size: 25,
           offset: 5,
         },
-        options: { signal: controller.signal },
       },
     ]);
   });
@@ -239,8 +235,7 @@ describe("ApiClient task search", () => {
       code: -32052,
       method: "workflow.task.search",
     });
-    expect(transport.calls).toEqual([]);
-    expect(transport.dedicatedCalls).toEqual([
+    expect(transport.calls).toEqual([
       {
         method: "workflow.task.search",
         params: {
