+++ apps/desktop/src/app-facade/TaskSearchMemoryProvider.tsx
@@ -5,40 +5,29 @@ import {
   type TaskSearchMemory,
   type TaskSearchMemorySelection,
 } from "./taskSearchMemoryContext";
-import type { TaskSearchScope } from "./taskSearchScope";
 
 export function TaskSearchMemoryProvider({ children }: Readonly<{ children: ReactNode }>) {
   const [query, setQuery] = useState("");
-  const [globalSelection, setGlobalSelection] = useState<TaskSearchMemorySelection | null>(null);
-  const [projectSelections, setProjectSelections] =
-    useState<ReadonlyMap<string, TaskSearchMemorySelection>>(() => new Map());
+  const [selections, setSelections] = useState<ReadonlyMap<string | null, TaskSearchMemorySelection>>(
+    () => new Map(),
+  );
   const rememberSelection = useCallback(
     (nextSelection: TaskSearchMemorySelection): void => {
-      if (nextSelection.scope.kind === "global") {
-        setGlobalSelection((current) =>
-          current?.key === nextSelection.key && current.query === nextSelection.query
-            ? current
-            : nextSelection,
-        );
-        return;
-      }
-      const projectID = nextSelection.scope.projectID;
-      setProjectSelections((current) => {
-        const previous = current.get(projectID);
+      setSelections((current) => {
+        const previous = current.get(nextSelection.projectID);
         if (previous?.key === nextSelection.key && previous.query === nextSelection.query) {
           return current;
         }
         const next = new Map(current);
-        next.set(projectID, nextSelection);
+        next.set(nextSelection.projectID, nextSelection);
         return next;
       });
     },
-    [],
+    [setSelections],
   );
   const selectionFor = useCallback(
-    (scope: TaskSearchScope): TaskSearchMemorySelection | null =>
-      scope.kind === "global" ? globalSelection : (projectSelections.get(scope.projectID) ?? null),
-    [globalSelection, projectSelections],
+    (projectID: string | null): TaskSearchMemorySelection | null => selections.get(projectID) ?? null,
+    [selections],
   );
   const memory = useMemo<TaskSearchMemory>(
     () => ({ query, rememberSelection, selectionFor, setQuery }),
