+++ apps/desktop/src/app-facade/index.ts
@@ -21,10 +21,8 @@ export * from "./statusStore";
 export * from "./taskDetailInitialFocus";
 export * from "./taskSearchMemoryContext";
 export * from "./taskSearchQueries";
-export * from "./taskSearchScope";
 export * from "./textFieldSubmitShortcut";
 export * from "./useAppServices";
-export * from "./useRetainedQueryData";
 export * from "./useConnectionSnapshot";
 export * from "./useNativeDialogFallback";
 export * from "./useStatusController";
@@ -32,4 +30,5 @@ export * from "./viewTransitions";
 export * from "./windowChromeTitle";
 export * from "./WindowChromeTitleProvider";
 export * from "./TaskSearchMemoryProvider";
+export { useRetainedQueryData } from "./useRetainedQueryData";
 export * from "./workflowProjectEvents";
