+++ apps/desktop/src/app-facade/taskSearchMemoryContext.ts
@@ -1,17 +1,15 @@
 import { createContext, useContext } from "react";
 
-import type { TaskSearchScope } from "./taskSearchScope";
-
 export type TaskSearchMemorySelection = Readonly<{
   key: string;
-  scope: TaskSearchScope;
+  projectID: string | null;
   query: string;
 }>;
 
 export type TaskSearchMemory = Readonly<{
   query: string;
   rememberSelection(selection: TaskSearchMemorySelection): void;
-  selectionFor(scope: TaskSearchScope): TaskSearchMemorySelection | null;
+  selectionFor(projectID: string | null): TaskSearchMemorySelection | null;
   setQuery(query: string): void;
 }>;
