--- apps/desktop/src/app-facade/taskSearchScope.ts
@@ -1,9 +1,0 @@
-export type TaskSearchScope =
-  | Readonly<{ kind: "global" }>
-  | Readonly<{ kind: "project"; projectID: string }>;
-
-export function taskSearchScopesEqual(left: TaskSearchScope, right: TaskSearchScope): boolean {
-  return left.kind === "global"
-    ? right.kind === "global"
-    : right.kind === "project" && left.projectID === right.projectID;
-}
