+++ apps/desktop/src/shared/task-search/index.ts
@@ -1,8 +1,1 @@
-export {
-  TaskSearchGlobalTrigger,
-  TaskSearchHost,
-  TaskSearchProjectTrigger,
-  TaskSearchProvider,
-  type TaskSearchActivationPolicy,
-  type TaskSearchInvocation,
-} from "./TaskSearch";
+export { BoardTaskSearchChrome } from "./TaskSearch";
