+++ apps/desktop/src/shared/task-search/taskSearchSelection.ts
@@ -1,11 +1,6 @@
 import { useCallback, useRef, useState } from "react";
 
-import {
-  taskSearchScopesEqual,
-  useTaskSearchMemory,
-  type TaskSearchMemorySelection,
-  type TaskSearchScope,
-} from "@/app-facade";
+import { useTaskSearchMemory, type TaskSearchMemorySelection } from "@/app-facade";
 import { prefersReducedMotion } from "@/ui";
 import type { TaskSearchResultItem as SearchResult } from "./TaskSearchResult";
 
@@ -16,24 +11,24 @@ export type TaskSearchScrollRequest = Readonly<{
 }>;
 
 export function useTaskSearchSelection(
-  scope: TaskSearchScope,
+  projectID: string | null,
   displayedQuery: string | null,
   results: readonly SearchResult[],
 ) {
   const memory = useTaskSearchMemory();
-  const selected = memory.selectionFor(scope);
+  const selected = memory.selectionFor(projectID);
   const [scrollRequest, setScrollRequest] = useState<TaskSearchScrollRequest | null>(null);
   const nextScrollRequestIDRef = useRef(1);
-  const activeKey = resolveActiveSearchResult(results, selected, scope, displayedQuery);
+  const activeKey = resolveActiveSearchResult(results, selected, projectID, displayedQuery);
   const select = useCallback(
     (key: string): void => {
       if (displayedQuery === null) {
         throw new Error("Task Search result selection requires a displayed query.");
       }
-      const next = { key, scope, query: displayedQuery };
+      const next = { key, projectID, query: displayedQuery };
       memory.rememberSelection(next);
     },
-    [displayedQuery, memory, scope],
+    [displayedQuery, memory, projectID],
   );
   const requestReveal = useCallback(
     (key: string, behavior: "auto" | "smooth"): void => {
@@ -97,15 +92,10 @@ export function adjacentSearchResult(
 function resolveActiveSearchResult(
   results: readonly SearchResult[],
   selected: TaskSearchMemorySelection | null,
-  scope: TaskSearchScope,
+  projectID: string | null,
   displayedQuery: string | null,
 ): string | null {
-  if (
-    displayedQuery !== null &&
-    selected !== null &&
-    taskSearchScopesEqual(selected.scope, scope) &&
-    selected.query === displayedQuery
-  ) {
+  if (displayedQuery !== null && selected?.projectID === projectID && selected.query === displayedQuery) {
     return results.some((result) => result.key === selected.key) ? selected.key : (results[0]?.key ?? null);
   }
   return results[0]?.key ?? null;
